normalize pid over nanos (the robot will go nuts since nothing is tuned)

+++ TeamCode/src/main/java/com/kuriosityrobotics/firstforward/robot/util/PID/ClassicalPID.java
@@ -10,8 +10,7 @@ public class ClassicalPID {
     private double lastError;
     private double errorSum;
     private double errorChange;
-
-    private long lastUpdateTime;
+    private double lastUpdateTime;
 
     /**
      * Constructs a ClassicalPIDController
